+++ app/models/user.rb
@@ -14,6 +14,10 @@ class User < ApplicationRecord
            through: :service_user_relationships,
            source: :service,
            class_name: "Service"
+  has_many :user_categories, dependent: :destroy
+  has_many :categories, through: :user_categories
+  has_many :user_research_areas, dependent: :destroy
+  has_many :research_areas, through: :user_research_areas
 
   validates :first_name, presence: true
   validates :last_name, presence: true